requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
python-multipart==0.0.9
pytest==8.3.4
httpx==0.27.2
python-dotenv==1.0.1
scikit-learn==1.5.2
